add: alex api

+++ app/src/main/java/com/example/rate_your_professor/model/pageList/TeacherInfo.kt
@@ -3,7 +3,7 @@ package com.example.rate_your_professor.model.pageList
 data class TeacherInfo(
         var teacherName: String,
         var ratting: Float,
-        var wouldTakeAgainPercen: Int,
+        var wouldTakeAgainPercen: Float,
         var levelOfDifficulty: Float,
         var startColour: String = ""
 )
